fix: folder creation bug in export_video_data()
Fix issue causing extra member video data folders to be created due to
'member_name' reformatting taking place after checking if the folder
path exists when it should be before the check. This resulted in a
second folder being created for each member that does not follow the
correct naming format.