feat(wip): improve functionality for summarize_content_types()
Improve the function by having it consider all content types and issue
a count of zero for content types that do not appear in the data. The
previous version did not do this, resulting in missing results.

Also, temporarily list all possible content types as a constant in the
module. They will be moved to a centralized location later.