feat(upgrade): improve functionality of summarize_video_data()
Modify the function to output a DataFrame that uses multilevel columns
instead of a normal DataFrame. This is due to the large number of
columns that will be in the DataFrame, which increases the need for
better organization. To allow this new structure, the way results are
stored as they come in and how the DataFrame is created are modified
as well.

Also, change 'summary' to 'member_summary' to improve clarity and
avoid confusion with other 'summary' variables.