feat(upgrade): improve functionality of summarize_video_types()
Improve the function by having it consider all video types, not just
the video types present in the data. For video types not present, issue
a count of zero. This change is due to results being missing if a video
type is not present in the data. For example, video data without any
Shorts will be missing a 'short_(count)' value.