feat(wip): add summarize_publish_datetimes()
Replace the placeholder and add code to summarize publish datetimes
of videos, e.g. the number of videos published in a given year. The
function has the option to also summarize data for just live streams,
e.g. the number of live streams in a given year, in addition to
summarizing the data as a whole.

Also, add placeholders for helper functions that will count the number
of videos published on a monthly and yearly basis.