feat(upgrade): improve functionality of summary_stats()
Add support for data stored as Timedeltas. round() does not accept
Timedeltas, so rounding results is no longer always done. As a further
upgrade, make rounding and the number of decimal places to use be up
to the user.

Also, use pd.isna() to check for NaN values instead of math.isnan().
The latter is fast but too limited. It requires real numbers and does
not accept Timedeltas.