feat(upgrade): modify formatting of video summary data
Modify how summarize_video_data() stores and presents member names.

Previously, the column containing member names had 'member_name' as the
Level 0 column name but lacked a Level 1 column name. This resulted in
Pandas assigning a name when the dataset is imported as a DataFrame.

Instead, provide column names for both levels to prevent this from
occurring and to ensure that the formatting of the DataFrame is
consistent regardless if the data was generated in the current session
or imported from a previous session.