feat(wip): update visualization prototyping notebook
Add a function, total_video_duration(), to plot the total video
duration of individual members while giving the user the option to
specify what unit of time to use, e.g. days, if the results are to be
sorted, and how to sort them.

Also, run the function and include some sample results.