feat(upgrade): improve importing of video data
Improve the importing of video data to include the date and time when
each dataset was created. This is done so that the date and time can be
easily referenced and used during data visualization.

Also, modify the process to more closely match its channel data
equivalent, improving consistency and maintainability.