feat(upgrade): improve data exporting
Improve export_channel_data() and export_video_data() by requiring the
functions to always export data, instead of only when the 'export_data'
parameter is True.

Also, modify channel and video data collection functions that call the
functions above to be compatible with the above change and to check if
'export_data' is True before calling the functions.

This is done because checking the value of 'export_data' in the export
functions is unnecessary. If the functions are called, it is expected
that the data is meant to be exported. It makes more sense to check the
value before calling an export function.