feat(upgrade): improve channel data collection
Improve channel data collection functions to include the date and time
when the data was obtained. This allows collected data to follow the
same structure as imported data.

Also, adjust the formatting to be consistent across similar functions,
and rename 'data' variables to be more specific, e.g. 'channel_data'.