fix: resolve video data collection bug
Resolve a bug in get_video_ids() that prevents members with no videos
from being skipped, leading to folder(s) and files being created that
contain no data and should not exist.

The bug is caused by the function not being updated to be compatible
with changes made to the video data collection process, leading to the
wrong object being referenced.