fix: resolve channel data importing bug
Resolve a bug that causes the oldest data in a session to always be
imported instead of the newest data.

For sessions where there is only one version of a dataset, the bug does
not cause problems. However, for sessions where there are multiple
versions of a dataset, the wrong version is imported.