refactor: rewrite channel_stats_correlation() helper
Rewrite the code that creates and formats plots. Core functionality is
unchanged. However, the function now contains minimal repetition and
allows for changes to the plots to be made more easily, e.g. A
formatting change for all x-axis labels can be done with one change to
the code instead of three.