feat(upgrade): Improve plot and colour data merging
Modify how merge_colour_data() merges datasets so that members not
present in the colour dataset are still included in the final dataset.
This resolves an issue where some members and their data may be missing
on plots.

Also, automatically assign a colour for members not present in the
colour dataset. This colour can be overridden by the user.