refactor: Update summarizing by unit to use session settings
Update the process of summarizing data by unit to use group, branch,
and unit names that are automatically extracted and stored in session
settings instead of referencing names that are hard-coded into the
module and need to be manually updated over time, saving time and
reducing potential issues in the long term.

Also, delete the hard-coded names because they are no longer required.